Created "database" package and moved "DatabaseController" class to it. Created "networking.messages" package and moved "ClientAuthenticationRequest", "ClientAuthenticationResponse" and "Message" classes (from "networking" package to it. Minor changes to other classes due to refactoring package paths.
