+++ src/database/DatabaseController.java
@@ -1,4 +1,4 @@
-package networking;
+package database;
 
 import java.sql.*;
 import java.sql.Connection;
