+++ src/networking/clientside/ClientThread.java
@@ -6,8 +6,8 @@ import java.io.ObjectOutputStream;
 import java.net.Socket;
 import javax.net.SocketFactory;
 import javax.net.ssl.SSLContext;
-import networking.ClientAuthenticationRequest;
-import networking.ClientAuthenticationResponse;
+import networking.messages.ClientAuthenticationRequest;
+import networking.messages.ClientAuthenticationResponse;
 import networking.Connection;
 import networking.MessageHandler;
