+++ src/networking/messages/ClientAuthenticationRequest.java
@@ -1,4 +1,4 @@
-package networking;
+package networking.messages;
 
 /**
  *
