+++ src/networking/messages/ClientAuthenticationResponse.java
@@ -1,4 +1,4 @@
-package networking;
+package networking.messages;
 
 /**
  *
